requirements.txt





flask
werkzeug
requests
gunicorn
psycopg2-binary
flask-cors
